requirements.txt
requests
pyyaml
schedule
python-dateutil
pprint
colorama